refactor: Clean up and organize DTOs for employee, event participation, login, office attendance, and room bookings
