+++ src/server/Controllers/RoomBookingsController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Calender_WebApp.Models;
 using Calender_WebApp.Services.Interfaces;
+using Calender_WebApp.Dtos;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Calender_WebApp.Controllers;
@@ -69,44 +70,6 @@ public class RoomBookingsController : ControllerBase
 		{
 			return Conflict(ex.Message);
 		}
-	}
-
-	/// <summary>
-	/// Request model for updating booking start time. Requires full composite key to identify existing booking.
-	/// </summary>
-	public class UpdateStartTimeRequest
-	{
-		public int RoomId { get; set; }
-		public int UserId { get; set; }
-		public DateTime BookingDate { get; set; }
-		public TimeSpan StartTime { get; set; }
-		public TimeSpan EndTime { get; set; }
-		public TimeSpan NewStartTime { get; set; }
-	}
-
-	/// <summary>
-	/// Request model for updating booking end time. Requires full composite key to identify existing booking.
-	/// </summary>
-	public class UpdateEndTimeRequest
-	{
-		public int RoomId { get; set; }
-		public int UserId { get; set; }
-		public DateTime BookingDate { get; set; }
-		public TimeSpan StartTime { get; set; }
-		public TimeSpan EndTime { get; set; }
-		public TimeSpan NewEndTime { get; set; }
-	}
-
-	/// <summary>
-	/// Request model for deleting room booking. Uses composite key instead of single booking ID.
-	/// </summary>
-	public class DeleteBookingRequest
-	{
-		public int RoomId { get; set; }
-		public int UserId { get; set; }
-		public DateTime BookingDate { get; set; }
-		public TimeSpan StartTime { get; set; }
-		public TimeSpan EndTime { get; set; }
 	}
 
 	[HttpPut("{bookingid}")]
